🚧 Wip: Keep list of Ingredient(-objects)
Afte shopping list was edited, one has only string and the information
and mechanics of the Ingredient-class were lost. Now, after parsing the
shopping_list after editing, the initial 'all_ingredients' list is
filtered to keep the remaining ones.

Due to rebuild, adding URLs is currently broken and will be fixed in the
following commit(s).