🎨 Update: Remove 'category_weight'
category_weight was removed because this causes too much trouble during
the handling because each category_weight has to be added in a second
CSV file. It is also laborious to explain the point behind
'category_weight' given that it has no/lost its usecase. The program aims
to do the grocery shopping at home via the browser and not in the
store. (With 'category_weights') it was possible to sort the
categories (and hence the ingredients) according to their lineup in the
supermarket to minimize walking.)